Extracted Options interface.
