+++ src/core/NearWalletSelector.tsx
@@ -5,26 +5,11 @@ import WalletController from "../controllers/WalletController";
 import Contract from "./Contract";
 import Modal from "../modal/Modal";
 import EventHandler, { Emitter, EventList } from "../utils/EventsHandler";
-import getConfig, { NetworkId } from "../config";
+import getConfig from "../config";
 import ProviderService from "../services/provider/ProviderService";
 import { updateState } from "../state/State";
 import { MODAL_ELEMENT_ID } from "../constants";
-
-export type BuiltInWalletId = "near-wallet" | "sender-wallet" | "ledger-wallet";
-export type UITheme = "dark" | "light";
-
-export interface Options {
-  wallets: Array<BuiltInWalletId>;
-  networkId: NetworkId;
-  contract: {
-    contractId: string;
-    methodNames?: Array<string>;
-  };
-  ui?: {
-    theme?: "dark" | "light";
-    description?: string;
-  };
-}
+import { BuiltInWalletId, Options } from "../interfaces/Options";
 
 export default class NearWalletSelector {
   private options: Options;
