+++ src/wallets/browser/NearWallet.ts
@@ -1,7 +1,7 @@
 import { WalletConnection, connect, keyStores } from "near-api-js";
 
 import getConfig from "../../config";
-import { Options } from "../../core/NearWalletSelector";
+import { Options } from "../../interfaces/Options";
 import { Emitter } from "../../utils/EventsHandler";
 import { logger } from "../../services/logging.service";
 import { transformActions } from "../actions";
